[API-93] Split HA operations into service-call/state/history; index.ts becomes a barrel.
